+++ .eslintrc.cjs
@@ -97,7 +97,7 @@ module.exports = {
         ],
         '@typescript-eslint/prefer-optional-chain': 'warn',
         'camelcase': 'off',
-        'max-len': ['warn', { code: 120, comments: null }],
+        'max-len': ['warn', { code: 120, comments: 120 }],
         '@typescript-eslint/naming-convention': [
           'error',
           // {
